requirements_all.txt
PyYAML
python-docx